Refactor password entry component
